requirements.txt
# venv\Scripts\activate <---- To activate the virtual environment on Windows
# source venv/bin/activate <---- To activate the virtual environment on macOS/Linux
# pip install -r requirements.txt
numpy
pandas
matplotlib
scipy
faiss-cpu  # or faiss-gpu if you have CUDA support (Later)
networkx